.nuspecs added for Security module
