+++ Example/Example/Example.csproj
@@ -112,6 +112,6 @@
   <ItemGroup>
     <ProjectReference Include="..\..\Core\Core\Core.csproj" />
     <ProjectReference Include="..\..\Auth\Auth\Auth.csproj" />
-    <ProjectReference Include="..\..\Security\Security.csproj" />
+    <ProjectReference Include="..\..\Security\Security\Security.csproj" />
   </ItemGroup>
 </Project>
