+++ Security/Security.Platform.Android/Checks/EncryptionCheck.cs
