+++ Security/Security.Platform.Android/Checks/NoDebuggerCheck.cs
