+++ Security/Security.Platform.Android/Checks/NotInEmulatorCheck.cs
