+++ Security/Security.Platform.Android/Checks/ScreenLockCheck.cs
