+++ Security/Security.Platform.Android/SecurityService.cs
