+++ Security/Security.Platform.iOS/Checks/DeviceLockCheck.cs
