+++ Security/Security.Platform.iOS/Checks/NoDebuggerCheck.cs
