+++ Security/Security.Platform.iOS/Checks/NotInEmulatorCheck.cs
