+++ Security/Security/AbstractSecurityCheck.cs
